\name{Comorbidity_calculation}
\alias{Comorbidity_calculation}
\title{Compute Charlson, Elixhauser and combined comorbidity scores}

\description{
This script processes patient data containing ICD-10 codes to compute comorbidity scores using the Elixhauser and Charlson methods. The calculated scores are then used to create a coincidence matrix and visualize comorbidity relationships.

}

\examples{
# Process and calculate frailty scores
source("R/Comorbidity_calculation.R")
}
