% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity-package.R
\docType{package}
\name{LABTNSCPSS-package}
\alias{LABTNSCPSS-package}
\title{LABTNSCPSS: Calculation of Comorbidity and Frailty Scores}
\description{
Computes comorbidity indices and combined scores for different versions of ICD, including ICD-10-CA, ICD-10-CM, and ICD-11.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bayaniazadeh/LabTNSCPSSPackage}
}

}
\author{
\strong{Maintainer}: Azadeh Bayani \email{azadeh.bayani@umontreal.ca} (\href{https://orcid.org/0009-0002-7707-9602}{ORCID})

Other contributors:
\itemize{
  \item Jean Noel Nikiema \email{jean.nikiema@umontreal.ca} [contributor]
  \item Michèle Bally \email{michele.bally.chum@ssss.gouv.qc.ca} [contributor]
}

}
\keyword{internal}
