% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_PACKAGE.R
\docType{_PACKAGE}
\name{LABTNSCPSS}
\alias{LABTNSCPSS}
\title{LABTNSCPSS: Comorbidity and Frailty Score Toolkit}
\description{
This package provides tools for calculating comorbidity indices
and frailty scores from different ICD coding systems (ICD-10-CA, ICD-10-CM, ICD-11).
}
